fix BX_UNBOXING_IMMEDIATELY_REBOXED
git-svn-id: https://svn.exedio.com/cope/cope/trunk@24973 e7d4fc99-c606-0410-b9bf-843393a9eab7

+++ runtime/src/com/exedio/cope/HsqldbDialect.java
@@ -236,7 +236,7 @@ final class HsqldbDialect extends Dialect
 					final Object o = resultSet.getObject(1);
 					if(o==null)
 						throw new RuntimeException("null in sequence " + name);
-					return ((Integer)o).intValue();
+					return (Integer)o;
 				}
 			});
 		}
